Move inject from internal/ to pkg/
